+++ test/integration/tests/abrmd_policypassword.sh
@@ -47,7 +47,7 @@ tpm2_encryptdecrypt -c $key_ctx -o $encrypted_txt -i $plain_txt -p $testpswd
 
 tpm2_startauthsession --policy-session -S $session_ctx
 tpm2_policypassword -S $session_ctx -o $policypassword
-tpm2_encryptdecrypt -c $key_ctx -i $encrypted_txt -o $decrypted_txt -D \
+tpm2_encryptdecrypt -c $key_ctx -i $encrypted_txt -o $decrypted_txt -d \
   -p session:$session_ctx+$testpswd
 tpm2_flushcontext -S $session_ctx
 rm $session_ctx