+++ test/integration/tests/encryptdecrypt.sh
@@ -47,14 +47,14 @@ tpm2_load -Q -C primary.ctx -u key.pub -r key.priv -n key.name -o decrypt.ctx
 
 tpm2_encryptdecrypt -Q -c decrypt.ctx  -i secret.dat -o encrypt.out
 
-tpm2_encryptdecrypt -Q -c decrypt.ctx -D -i encrypt.out -o decrypt.out
+tpm2_encryptdecrypt -Q -c decrypt.ctx -d -i encrypt.out -o decrypt.out
 
 # Test using stdin/stdout
-cat secret.dat | tpm2_encryptdecrypt -c decrypt.ctx | tpm2_encryptdecrypt -c decrypt.ctx -D > secret2.dat
+cat secret.dat | tpm2_encryptdecrypt -c decrypt.ctx | tpm2_encryptdecrypt -c decrypt.ctx -d > secret2.dat
 
 # test using IVs
 dd if=/dev/urandom of=iv.dat bs=16 count=1
-cat secret.dat | tpm2_encryptdecrypt -c decrypt.ctx --iv iv.dat | tpm2_encryptdecrypt -c decrypt.ctx --iv iv.dat:iv2.dat -D > secret2.dat
+cat secret.dat | tpm2_encryptdecrypt -c decrypt.ctx --iv iv.dat | tpm2_encryptdecrypt -c decrypt.ctx --iv iv.dat:iv2.dat -d > secret2.dat
 
 cmp secret.dat secret2.dat
 
@@ -72,7 +72,7 @@ echo -n "1234567812345678" > secret.dat
 tpm2_encryptdecrypt -Q -c decrypt.ctx -G cbc -i secret.dat --iv=iv.dat  -o encrypt.out
 
 # Unspecified mode (figure out via readpublic)
-tpm2_encryptdecrypt -Q -D -c decrypt.ctx -i encrypt.out --iv iv.dat -o decrypt.out
+tpm2_encryptdecrypt -Q -d -c decrypt.ctx -i encrypt.out --iv iv.dat -o decrypt.out
 
 cmp secret.dat decrypt.out
 
@@ -80,8 +80,8 @@ cmp secret.dat decrypt.out
 tpm2_encryptdecrypt -Q -c decrypt.ctx -G cbc -i secret.dat --iv=iv.dat:iv2.dat -o encrypt.out
 tpm2_encryptdecrypt -Q -c decrypt.ctx -G cbc -i secret.dat --iv=iv2.dat -o encrypt2.out
 
-tpm2_encryptdecrypt -Q -D -c decrypt.ctx -i encrypt.out --iv iv.dat -o decrypt.out
-tpm2_encryptdecrypt -Q -D -c decrypt.ctx -i encrypt2.out --iv iv2.dat -o decrypt2.out
+tpm2_encryptdecrypt -Q -d -c decrypt.ctx -i encrypt.out --iv iv.dat -o decrypt.out
+tpm2_encryptdecrypt -Q -d -c decrypt.ctx -i encrypt2.out --iv iv2.dat -o decrypt2.out
 
 cmp secret.dat decrypt.out
 cmp secret.dat decrypt2.out
@@ -89,14 +89,14 @@ cmp secret.dat decrypt2.out
 # Test that input data sizes greater than TPM2_MAX_BUFFER or 1024 work
 dd if=/dev/zero bs=1 count=2048 status=none of=secret2.dat
 cat secret2.dat | tpm2_encryptdecrypt -Q -c decrypt.ctx -o encrypt.out
-tpm2_encryptdecrypt -Q -c decrypt.ctx -i encrypt.out -D -o decrypt.out
+tpm2_encryptdecrypt -Q -c decrypt.ctx -i encrypt.out -d -o decrypt.out
 
 cmp secret2.dat decrypt.out
 
 # Test that last block in input data shorter than block length has pkcs7 padding
 dd if=/dev/zero bs=1 count=2050 status=none of=secret2.dat
 cat secret2.dat | tpm2_encryptdecrypt -Q -c decrypt.ctx -o encrypt.out -e
-tpm2_encryptdecrypt -Q -c decrypt.ctx -i encrypt.out -D -o decrypt.out
+tpm2_encryptdecrypt -Q -c decrypt.ctx -i encrypt.out -d -o decrypt.out
 ## Last block is short 14 or hex 0E trailing bytes
 echo "0e0e0e0e0e0e0e0e0e0e0e0e0e0e" | xxd -r -p >> secret2.dat
 cmp secret2.dat decrypt.out
@@ -104,7 +104,7 @@ cmp secret2.dat decrypt.out
 # Test that pkcs7 padding is added as last block for block length aligned inputs
 dd if=/dev/zero bs=1 count=2048 status=none of=secret2.dat
 cat secret2.dat | tpm2_encryptdecrypt -Q -c decrypt.ctx -o encrypt.out -e
-tpm2_encryptdecrypt -Q -c decrypt.ctx -i encrypt.out -D -o decrypt.out
+tpm2_encryptdecrypt -Q -c decrypt.ctx -i encrypt.out -d -o decrypt.out
 ## Last block is short 14 or hex 0E trailing bytes
 echo "10101010101010101010101010101010" | xxd -r -p >> secret2.dat
 cmp secret2.dat decrypt.out
@@ -112,13 +112,13 @@ cmp secret2.dat decrypt.out
 # Test pkcs7 padding is stripped from input data is shorter than block length
 dd if=/dev/zero bs=1 count=2050 status=none of=secret2.dat
 cat secret2.dat | tpm2_encryptdecrypt -Q -c decrypt.ctx -o encrypt.out -e
-tpm2_encryptdecrypt -Q -c decrypt.ctx -i encrypt.out -D -o decrypt.out -e
+tpm2_encryptdecrypt -Q -c decrypt.ctx -i encrypt.out -d -o decrypt.out -e
 cmp secret2.dat decrypt.out
 
 # Test that pkcs7 pad is stripped off last block for block length aligned inputs
 dd if=/dev/zero bs=1 count=2048 status=none of=secret2.dat
 cat secret2.dat | tpm2_encryptdecrypt -Q -c decrypt.ctx -o encrypt.out -e
-tpm2_encryptdecrypt -Q -c decrypt.ctx -i encrypt.out -D -o decrypt.out -e
+tpm2_encryptdecrypt -Q -c decrypt.ctx -i encrypt.out -d -o decrypt.out -e
 cmp secret2.dat decrypt.out
 
 # Negative that bad mode fails